Fix programs regarding duplication and Drop.
In #346 for #345, I removed the destroy_* methods from the backend
interface to allow backends to be more flexible about when resources
should be dropped. That had the side-effect that duplication of
`Program` in the luminance-gl backend would cause a non-wanted drop of a
GPU scarce resource (and potentially double-free).

This commit fixes that.