[luminance-webgl] Fix index buffer creation on WebGL2.
When binding a buffer for the first time, on WebGL2, the buffer’s
internal representation is tied to the target argument we use when
binding. This result in issues if we try to create and store data as
ARRAY_BUFFER first then use ELEMENT_ARRAY_BUFFER later.

This commit fixes the creation of index buffers.