Refactor render-{vertex,instance}-nb.
This commit refactors the usual way of getting the default number of
vertices and vertex instances to render. Prior to this commit, this was
encoded by `Tess::vert_nb` and `Tess::inst_nb`, and was provided by the
backend implementation.

The new situation uses `Tess::vert_nb`, `Tess::idx_nb` and
`Tess::inst_nb` to get the exact number of elements in the storage. The
introduction of `Tess::idx_nb` relates to #478. The behavior of how the
default numbers are guessed doesn’t change but new functions were
introduced to query the “render” versions:

- `Tess::render_vert_nb` to query the number of vertices that will be
  rendered by default. This is effectively either the number of vertices
  or indices, depending which kind of geometry was made.
- `Tess::render_inst_nb` to query the number of vertex instances that
  will be rendered by default.