Add premises of the Query API.
This commit introduces the very beginning of the Query API, that will be
used to implement various queries on the backend to provide versions,
names and capabilities of the GPU / hardware.

This commit introduces the main API along with:

- The backend “author” (vendor, company, etc.).
- The backend name.
- The backend version.
- The backend shading language version.

This commit holds both the implementations for the OpenGL 3.3 and WebGL2
backends.

Relates to #492.