[luminance-webgl] Fix buffer target corruption in WebGL2 backend.
In the OpenGL 3.3 backend, there is an optimization that allows a
buffer object to be created and bound to a given target (for instance
`ARRAY_BUFFER`) and allows to rebind the buffer later to another
target (for instance `ELEMENT_ARRAY_BUFFER`). That optimization is handy
because it allows to dissociate data storage from data representation,
allowing to send a bunch of bytes and then using the target as source of
representation.

Well, turns out that optimization doesn’t work in WebGL2, so this commit
implements a way to keep track of what a buffer was created as and then
re-use that information every time we need to bind the buffer. That
should fix #483 (and probably other bugs no one has encountered yet).

Some integration tests are needed to ensure the fix does what it should.