Migrate hello-world to the platform-agnostic arch.
First of a series of refactoring, this commit introduces the concept of
`Example`, a trait that can be implemented to provide examples agnostic
of the platform they will be executed on. To showcase how to use it, the
`hello-world` example was migrated to this new architecture.

This commit doesn’t allow to run the examples yet. The next commit will
implement the _desktop_ and _web_ platform to demonstrate how to do it.