Remove Buffer from the public API.
Buffers are not really needed / used by people and are an implementation
detail, as they are required for (currently) three things:

- Vertex buffers (for `Tess`).
- Index buffers (still for `Tess`).
- And uniform buffers, which are known to be problematic because of the
  unsafe nature of using them and will require a redesign.

So better get rid of the concept of buffers and expose directly the
upper abstractions, such as slicing `Tess`, which is already supported
and then won’t be a breaking change. The only case that breaks is when
people want to build a `Tess` from GPU buffers, which is not going to be
supported anymore.

Relates to #483.