Change the Uniformable backend interface.
The implementor and the type variable are now swapped. Instead of
implementing it as:

```rust
impl Uniformable<TheBackendType> for [f32; 3] { /* … */ }
```

It is now implemented as:

```rust
impl Uniformable<[f32; 3]> for TheBackendType { /* … */ }
```

This allows for better polymorphic code. Next commit should introduce a
type family to be able to manipulate references correctly, and use the
`luminance::shader::types` types, since the current implementors are
wrong and ambiguous.

Relates to #438.